% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csv.quibl.R
\name{read.csv.quibl}
\alias{read.csv.quibl}
\title{Read output file from QuIBL}
\usage{
read.csv.quibl(filepath)
}
\arguments{
\item{filepath}{QuIBL output file}
}
\description{
This function reads in the basic output generated by a QuIBL run
}
\examples{
read.csv.quibl(quiblFile.csv)
}
